refactor(subway): regroup classes to packages
