+++ src/main/java/subway/Application.java
@@ -4,6 +4,7 @@ import subway.domain.Line;
 import subway.domain.LineRepository;
 import subway.domain.Station;
 import subway.domain.StationRepository;
+import subway.menu.MainMenu;
 
 import java.util.ArrayList;
 import java.util.Arrays;
