+++ src/main/java/subway/menu/MainMenu.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.menu;
 
 import java.util.ArrayList;
 import java.util.Arrays;
